add CategoryFilter component with styled-components for improved UI
